9.27

+++ kindergarten-commons/kindergarten-common/src/main/java/com/zc/kindergarten/common/util/TreeNode.java
@@ -8,8 +8,8 @@ import java.util.List;
  * @create 2018/9/19.
  */
 public class TreeNode {
-    protected int id;
-    protected int parentId;
+    protected Long id;
+    protected Long parentId;
 
     List<TreeNode> children = new ArrayList<TreeNode>();
 
@@ -21,19 +21,19 @@ public class TreeNode {
         this.children = children;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getParentId() {
+    public Long getParentId() {
         return parentId;
     }
 
-    public void setParentId(int parentId) {
+    public void setParentId(Long parentId) {
         this.parentId = parentId;
     }
